requirements.txt
numpy
matplotlib
pandas
joblib
gym
numba
seaborn
pytest
PyOpenGL
PyOpenGL_accelerate